Factor GenFacades into a core library and a console application
