--- src/GenFacades.Desktop/project.json
@@ -1,22 +1,0 @@
-{
-  "dependencies": {
-    "Microsoft.Cci": "4.0.0-rc4-24217-00",
-    "System.Collections": "4.0.10",
-    "System.Console": "4.0.0",
-    "System.Diagnostics.Debug": "4.0.10",
-    "System.Diagnostics.FileVersionInfo": "4.0.0",
-    "System.Diagnostics.TextWriterTraceListener": "4.0.0",
-    "System.Diagnostics.TraceSource": "4.0.0",
-    "System.IO.FileSystem": "4.0.0",
-    "System.Linq": "4.0.0",
-    "System.Reflection": "4.0.10",
-    "System.Reflection.TypeExtensions": "4.0.0",
-    "System.Text.Encoding": "4.0.10"
-  },
-  "frameworks": {
-    "net46": {}
-  },
-  "runtimes": {
-    "win7-x64": {}
-  }
-}
