+++ src/GenFacades/GenFacades.Console/project.json
@@ -3,10 +3,10 @@
     "Microsoft.Cci": "4.0.0-rc4-24217-00",
     "System.Diagnostics.TextWriterTraceListener": "4.0.0",
     "System.Diagnostics.FileVersionInfo": "4.0.0",
-    "Microsoft.NETCore.App": {
-      "type": "platform",
-      "version": "1.0.0"
-    }
+    "System.Reflection.TypeExtensions": "4.1.0",
+    "System.Reflection.Extensions": "4.0.1",
+    "System.Linq": "4.1.0",
+    "System.Console": "4.0.0"
   },
   "frameworks": {
     "netcoreapp1.0": {
