+++ src/GenFacades/GenFacades.Core/FacadeGenerationException.cs
