+++ src/GenFacades/GenFacades.Core/VersionResourceSerializer.cs
