+++ src/GenFacades/GenFacades.Core/project.json
@@ -1,0 +1,15 @@
+{
+  "dependencies": {
+    "Microsoft.Cci": "4.0.0-rc4-24217-00",
+    "System.Diagnostics.TextWriterTraceListener": "4.0.0",
+    "System.Diagnostics.FileVersionInfo": "4.0.0",
+    "System.Reflection.TypeExtensions": "4.1.0",
+    "System.Reflection.Extensions": "4.0.1",
+    "System.Linq": "4.1.0",
+  },
+  "frameworks": {
+    "netstandard1.3": {
+      "imports": "dnxcore50"
+    }
+  }
+}
