+++ src/nuget/Microsoft.DotNet.BuildTools.nuspec
@@ -29,7 +29,8 @@
     <file src="run\run.exe" target="lib" />
     <file src="run\run.runtimeconfig.json" target="lib" />
     <file src="BclRewriter\BclRewriter.exe" target="lib" />
-    <file src="GenFacades\GenFacades.exe" target="lib" />
+    <file src="GenFacades.Console\GenFacades.exe" target="lib" />
+    <file src="GenFacades.Core\GenFacades.Core.dll" target="lib" />
     <file src="GenAPI\GenAPI.exe" target="lib" />
     <file src="ApiCompat\ApiCompat.exe" target="lib" />
     <file src="Microsoft.DotNet.CodeAnalysis\Microsoft.DotNet.CodeAnalysis.dll" target="lib\analyzers" />
